Skip linting in node_modules

+++ .buildkite/steps/lint.sh
@@ -5,6 +5,7 @@ grep -rl '^#!/.*sh' . | while read -r file ; do
   [[ $file =~ \.git ]] && continue
   [[ $file =~ init\.d ]] && continue
   [[ $file =~ vendor ]] && continue
+  [[ $file =~ node_modules ]] && continue
 
   echo "Processing $file"
   docker run --rm -v "$PWD:/mnt" koalaman/shellcheck "$file"
